* added launch storyboard to enable split view multitasking (works on iPad Air2 simulator) * added functions to handle width and height differences among different iOS devices * fixed some of the warnings (due to deprecated functions) * keyboard is shown at startup of the app * refactoring

+++ AmikoDesitin/MLProgressViewController.h
@@ -33,7 +33,7 @@
 
 - (void) setMessage:(NSString *)msg;
 - (void) start;
-- (void) updateWith:(int)downloadedBytes andWith:(int)expectedBytes;
+- (void) updateWith:(long)downloadedBytes andWith:(long)expectedBytes;
 - (void) remove;
 
 @end
